+++ docs/changelog.rst
@@ -98,6 +98,9 @@ To update |kitty|, :doc:`follow the instructions <binary>`.
 - Only update the selected text to clipboard when the selection is finished,
   not continuously as it is updated. (:iss:`1460`)
 
+- Allow setting :opt:`active_border_color` to ``none`` to not draw a border
+  around the active window (:iss:`805`)
+
 
 0.13.3 [2019-01-19]
 ------------------------------